Optimize module structure.
