+++ pom.xml
@@ -24,19 +24,51 @@
         <artifactId>spring-boot-starter-parent</artifactId>
         <version>2.4.10</version>
     </parent>
+
     <groupId>com.zergclan</groupId>
     <artifactId>wormhole</artifactId>
-    <version>${wormhole.sersion}</version>
+    <version>0.1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>${project.artifactId}</name>
+    <url>https://github.com/zergclan/wormhole</url>
+
+    <licenses>
+        <license>
+            <name>Apache License, Version 2.0</name>
+            <url>https://www.apache.org/licenses/LICENSE-2.0.txt</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+
+    <developers>
+        <developer>
+            <name>yanghou</name>
+            <email>jacky7boy@163.com</email>
+        </developer>
+        <developer>
+            <name>WtlKoma</name>
+            <email></email>
+        </developer>
+        <developer>
+            <name>wenguozhang</name>
+            <email></email>
+        </developer>
+        <developer>
+            <name>rabbit1225</name>
+            <email></email>
+        </developer>
+    </developers>
 
     <properties>
         <wormhole.sersion>0.1.0-SNAPSHOT</wormhole.sersion>
         <java.version>1.8</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <!-- tools -->
+        <guava.sersion>30.1.1-jre</guava.sersion>
+        <gson.sersion>2.8.9</gson.sersion>
         <disruptor.version>3.4.4</disruptor.version>
-        <mybatis-spring-boot-starter.version>2.1.4</mybatis-spring-boot-starter.version>
+        <lombok.version>1.18.20</lombok.version>
+        <h2.version>1.4.200</h2.version>
         <!-- test -->
         <junit5.version>5.7.2</junit5.version>
         <!-- maven-plugin -->
@@ -55,20 +87,41 @@
 
     <dependencyManagement>
         <dependencies>
+            <!-- tools -->
             <dependency>
-                <groupId>org.mybatis.spring.boot</groupId>
-                <artifactId>mybatis-spring-boot-starter</artifactId>
-                <version>${mybatis-spring-boot-starter.version}</version>
+                <groupId>com.google.guava</groupId>
+                <artifactId>guava</artifactId>
+                <version>${guava.sersion}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.google.code.gson</groupId>
+                <artifactId>gson</artifactId>
+                <version>${gson.sersion}</version>
             </dependency>
             <dependency>
                 <groupId>com.lmax</groupId>
                 <artifactId>disruptor</artifactId>
                 <version>${disruptor.version}</version>
+                <scope>provided</scope>
             </dependency>
+            <dependency>
+                <groupId>org.projectlombok</groupId>
+                <artifactId>lombok</artifactId>
+                <version>${lombok.version}</version>
+                <scope>provided</scope>
+            </dependency>
+            <dependency>
+                <groupId>com.h2database</groupId>
+                <artifactId>h2</artifactId>
+                <version>${h2.version}</version>
+                <scope>provided</scope>
+            </dependency>
+            <!-- test -->
             <dependency>
                 <groupId>org.junit.jupiter</groupId>
                 <artifactId>junit-jupiter</artifactId>
                 <version>${junit5.version}</version>
+                <scope>test</scope>
             </dependency>
         </dependencies>
     </dependencyManagement>
@@ -77,6 +130,11 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 
@@ -122,7 +180,6 @@
                     </execution>
                 </executions>
             </plugin>
-
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>cobertura-maven-plugin</artifactId>
