+++ wormhole-features/wormhole-web/pom.xml
@@ -6,6 +6,7 @@
     <parent>
         <groupId>com.zergclan</groupId>
         <artifactId>wormhole-features</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>wormhole-web</artifactId>
     <name>${project.artifactId}</name>
