+++ wormhole-infra/pom.xml
@@ -23,6 +23,7 @@
     <parent>
         <groupId>com.zergclan</groupId>
         <artifactId>wormhole</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>wormhole-infra</artifactId>
     <packaging>pom</packaging>
