+++ wormhole-infra/wormhole-bus/pom.xml
@@ -23,14 +23,14 @@
     <parent>
         <groupId>com.zergclan</groupId>
         <artifactId>wormhole-infra</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>wormhole-bus</artifactId>
+    <packaging>pom</packaging>
     <name>${project.artifactId}</name>
 
-    <dependencies>
-        <dependency>
-            <groupId>com.lmax</groupId>
-            <artifactId>disruptor</artifactId>
-        </dependency>
-    </dependencies>
+    <modules>
+        <module>wormhole-bus-api</module>
+        <module>wormhole-bus-memory</module>
+    </modules>
 </project>
