+++ wormhole-infra/wormhole-common/pom.xml
@@ -23,13 +23,7 @@
     <parent>
         <groupId>com.zergclan</groupId>
         <artifactId>wormhole-infra</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>wormhole-common</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter</artifactId>
-        </dependency>
-    </dependencies>
 </project>
