+++ wormhole-infra/wormhole-context/pom.xml
@@ -23,6 +23,7 @@
     <parent>
         <groupId>com.zergclan</groupId>
         <artifactId>wormhole-infra</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>wormhole-context</artifactId>
     <name>${project.artifactId}</name>
