+++ wormhole-infra/wormhole-repository/pom.xml
@@ -23,7 +23,14 @@
     <parent>
         <groupId>com.zergclan</groupId>
         <artifactId>wormhole-infra</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>wormhole-repository</artifactId>
+    <packaging>pom</packaging>
     <name>${project.artifactId}</name>
+
+    <modules>
+        <module>wormhole-repository-api</module>
+        <module>wormhole-repository-core</module>
+    </modules>
 </project>
