+++ wormhole-infra/wormhole-repository/wormhole-repository-api/pom.xml
@@ -21,10 +21,10 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
+        <artifactId>wormhole-repository</artifactId>
         <groupId>com.zergclan</groupId>
-        <artifactId>wormhole-infra</artifactId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
-    <artifactId>wormhole-pipeline</artifactId>
+    <artifactId>wormhole-repository-api</artifactId>
     <name>${project.artifactId}</name>
 </project>
