+++ wormhole-infra/wormhole-repository/wormhole-repository-core/pom.xml
@@ -21,10 +21,22 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
+        <artifactId>wormhole-repository</artifactId>
         <groupId>com.zergclan</groupId>
-        <artifactId>wormhole-infra</artifactId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
-    <artifactId>wormhole-pipeline</artifactId>
+    <artifactId>wormhole-repository-core</artifactId>
     <name>${project.artifactId}</name>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.zergclan</groupId>
+            <artifactId>wormhole-repository-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.h2database</groupId>
+            <artifactId>h2</artifactId>
+        </dependency>
+    </dependencies>
 </project>
