+++ wormhole-writer/wormhole-writer-mysql/pom.xml
@@ -23,6 +23,7 @@
     <parent>
         <groupId>com.zergclan</groupId>
         <artifactId>wormhole-writer</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>wormhole-writer-mysql</artifactId>
     <name>${project.artifactId}</name>
